commit new

+++ src/main/java/com/example/demo1/models/Product.java
@@ -95,7 +95,7 @@ public class Product {
         return price;
     }
 
-    public void setPrice(double price) {
+    public void setPrice(int price) {
         this.price = price;
     }
